refactor: move venv setup to end of scripts
Moved the virtual environment creation prompt to the end of the pc.sh and termux.sh scripts.
This ensures that the virtual environment is created and activated before the env.sh script is sourced, resolving potential path issues.

Also fixed a bug in setup.sh where env.sh was being called twice.